Change branches to use
Now, when pushing to: main/next/exp, then the following branches, on the
xnvme repository is used:

* dockerize-for-main
* dockerize-for-next
* dockerize-for-exp

And will deploy containers to ':main', ':next', ':exp'.

Currently, then the github-action for the xNVMe repository only uses
':next'. However, with this, then it is more convenient to experimet
with changes to how the docker-images are built.

Signed-off-by: Simon A. F. Lund <os@safl.dk>